Started implementing dialog functionality
Aim here was to create a simple say() function for an actor. The
parameters are the string you want the actor to say and the time
in milliseconds you want it to display for.

This can be called from lua - in the demo the scarecrow speaks.

Note that this is just to get the basic idea of a dialog system -
as of yet the dialog is not rendered. There will be a modular speech
rendering system that will allow a developer to define how the
dialog in their game is displayed. For now, in place of this, a
printf function is used :)