Started a basic layout for the dialog frontend system.
The idea is that the game has a pluggable dialog frontend which is chosen
and/or defined by the user. Each frontend will derive from the DialogFrontend
base class.

When a character is speaking the scene takes their current line of dialog
and sends it to the current frontend for rendering.