Initial implementation of a rendering frontend for actors' dialog lines.
The default dialog frontend maps text objects to actors who are speaking,
and renders their lines. When an actor is finished speaking they are removed
from the map.

This is a rough draft. The dialog is rendered under the character's feet as
it's not clever enough yet to determine a position above actors' heads. Also
all dialog is rendered in the same font and color at the moment.